Add custom relay strategy
