+++ components/client-pangoro-s2s/src/feemarket.rs
@@ -1,0 +1,28 @@
+use messages_relay::message_lane::MessageLane;
+use messages_relay::message_lane_loop::{
+    SourceClient as MessageLaneSourceClient, TargetClient as MessageLaneTargetClient,
+};
+use messages_relay::relay_strategy::{RelayReference, RelayStrategy};
+
+#[derive(Clone)]
+pub struct PangoroRelayStrategy {}
+
+impl PangoroRelayStrategy {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
+
+#[async_trait::async_trait]
+impl RelayStrategy for PangoroRelayStrategy {
+    async fn decide<
+        P: MessageLane,
+        SourceClient: MessageLaneSourceClient<P>,
+        TargetClient: MessageLaneTargetClient<P>,
+    >(
+        &self,
+        reference: &mut RelayReference<P, SourceClient, TargetClient>,
+    ) -> bool {
+        true
+    }
+}
