+++ task/task-pangolin-pangoro/src/chains/pangoro.rs
@@ -152,7 +152,6 @@ mod s2s_messages {
     use frame_support::dispatch::GetDispatchInfo;
     use frame_support::weights::Weight;
     use messages_relay::message_lane::MessageLane;
-    use messages_relay::relay_strategy::AltruisticStrategy;
     use relay_substrate_client::{Client, IndexOf, TransactionSignScheme, UnsignedTransaction};
     use relay_utils::metrics::MetricsParams;
     use sp_core::{Bytes, Pair};
@@ -164,7 +163,7 @@ mod s2s_messages {
     use substrate_relay_helper::messages_target::SubstrateMessagesTarget;
 
     use component_pangolin_s2s::PangolinChain;
-    use component_pangoro_s2s::PangoroChain;
+    use component_pangoro_s2s::{PangoroChain, PangoroRelayStrategy};
 
     use crate::chains::pangolin::PangolinChainConst;
     use crate::chains::pangoro::PangoroChainConst;
@@ -323,7 +322,7 @@ mod s2s_messages {
                 <PangoroChainConst as ChainConst>::SigningParams,
                 PangolinChain,
                 <PangolinChainConst as ChainConst>::SigningParams,
-                AltruisticStrategy,
+                PangoroRelayStrategy,
             >,
         ) -> anyhow::Result<()> {
             let stall_timeout = Duration::from_secs(5 * 60);
